requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
pymongo==4.6.1
PyMySQL==1.1.0
google-generativeai==0.3.2
python-dotenv==1.0.0
